implements : extraction of square sets and solving them
changes : SInequality class
	1. changes floats to double
	2. improves documentation
	3. implements methods
		i.   getCoefficientRow()
		ii.  getVariableRow()
		iii. getConstantRow()
	4. improves toString() method

changes : Shipment class
	1. changes floats to doubles

changes : Main class
	1. implements methods
		i.  Solve(Vector<SInequality> sInequalities)
		ii. extractSquareSets(Vector<SInequality> similar_set, int cardinality,
				Vector<SInequality> buffer, int buffer_i, int input_i)
	2. improves documentation

+++ src/Physical/Shipment.java
@@ -29,20 +29,20 @@ public class Shipment {
     // Box fields
     // length >= breadth >= height > 0
     // weight > 0
-    float length;
-    float breadth;
-    float height;
-    float weight;
+    double length;
+    double breadth;
+    double height;
+    double weight;
 
     /**
      * Assert length, breadth, height > 0
      * Assert weight > 0
      */
-    public Shipment(String order_id, String id, float length, float breadth, float height, float weight) {
+    public Shipment(String order_id, String id, double length, double breadth, double height, double weight) {
         this.order_id = order_id;
         this.id = id;
 
-        float hbl[] = {
+        double hbl[] = {
                 height,
                 breadth,
                 length
